Refactored use-defs
