+++ external-crates/move/crates/move-analyzer/src/analysis/mod.rs
@@ -5,7 +5,8 @@ use crate::{
     compiler_info::CompilerInfo,
     symbols::{
         cursor::CursorContext,
-        types::{CompiledPkgInfo, SymbolsComputationData, UseDefMap},
+        types::{CompiledPkgInfo, SymbolsComputationData},
+        use_def::UseDefMap,
     },
     utils::expansion_mod_ident_to_map_key,
 };
