+++ external-crates/move/crates/move-analyzer/src/analysis/parsing_analysis.rs
@@ -7,10 +7,8 @@ use crate::{
         cursor::{CursorContext, CursorDefinition, CursorPosition},
         ignored_function, parsed_address, parsing_leading_and_mod_names_to_map_key,
         parsing_mod_def_to_map_key,
-        types::{
-            AutoImportInsertionInfo, AutoImportInsertionKind, CallInfo, DefMap, ModuleDefs,
-            References, UseDef, UseDefMap,
-        },
+        types::{AutoImportInsertionInfo, AutoImportInsertionKind, CallInfo, DefMap, ModuleDefs},
+        use_def::{References, UseDef, UseDefMap},
     },
     utils::{loc_end_to_lsp_position_opt, loc_start_to_lsp_position_opt},
 };
