+++ external-crates/move/crates/move-analyzer/src/symbols/requests.rs
@@ -7,7 +7,8 @@ use crate::{
     context::Context,
     symbols::{
         runner::SymbolicatorRunner,
-        types::{DefInfo, MemberDef, MemberDefInfo, ModuleDefs, Symbols, UseDef},
+        types::{DefInfo, MemberDef, MemberDefInfo, ModuleDefs, Symbols},
+        use_def::UseDef,
     },
     utils::lsp_position_to_loc,
 };
